Moves generic utils under utils folder.
