+++ ocean_provider/validation/requests.py
@@ -8,8 +8,8 @@ from flask_sieve.rules_processor import RulesProcessor
 from flask_sieve.validator import Validator
 from ocean_provider.exceptions import InvalidSignatureError
 from ocean_provider.user_nonce import get_nonce
-from ocean_provider.util import get_request_data
 from ocean_provider.utils.accounts import verify_signature
+from ocean_provider.utils.util import get_request_data
 
 
 class CustomJsonRequest(JsonRequest):
